pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"


[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = []
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
